--- a/src/wrapper/jnienv.rs
+++ b/src/wrapper/jnienv.rs
@@ -767,7 +767,7 @@ impl<'a> JNIEnv<'a> {
         class: T,
         method_id: U,
         ret: ReturnType,
-        args: &[JValue],
+        args: &[jvalue],
     ) -> Result<JValue<'a>>
     where
         T: Desc<'a, JClass<'c>>,
@@ -778,7 +778,6 @@ impl<'a> JNIEnv<'a> {
         let method_id = method_id.lookup(self)?.into_inner();
 
         let class = class.into_inner();
-        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
         let jni_args = args.as_ptr();
 
         // TODO clean this up
@@ -884,7 +883,7 @@ impl<'a> JNIEnv<'a> {
         obj: O,
         method_id: T,
         ret: ReturnType,
-        args: &[JValue],
+        args: &[jvalue],
     ) -> Result<JValue<'a>>
     where
         O: Into<JObject<'a>>,
@@ -894,7 +893,6 @@ impl<'a> JNIEnv<'a> {
 
         let obj = obj.into().into_inner();
 
-        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
         let jni_args = args.as_ptr();
 
         // TODO clean this up
@@ -981,7 +979,8 @@ impl<'a> JNIEnv<'a> {
 
         let class = self.auto_local(self.get_object_class(obj)?);
 
-        self.call_method_unchecked(obj, (&class, name, sig), parsed.ret, args)
+        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
+        self.call_method_unchecked(obj, (&class, name, sig), parsed.ret, &args)
     }
 
     /// Calls a static method safely. This comes with a number of
@@ -1016,7 +1015,8 @@ impl<'a> JNIEnv<'a> {
         // and we'll need that for the next call.
         let class = class.lookup(self)?;
 
-        self.call_static_method_unchecked(class, (class, name, sig), parsed.ret, args)
+        let args: Vec<jvalue> = args.iter().map(|v| v.to_jni()).collect();
+        self.call_static_method_unchecked(class, (class, name, sig), parsed.ret, &args)
     }
 
     /// Create a new object using a constructor. This is done safely using