## Overview

I've been depending on this branch that was shared as part of this issue https://github.com/jni-rs/jni-rs/issues/329 for some time and would be very interested in seeing these changes make it upstream, while it's not ideal that method calls currently need to dynamically allocate a vector on each call - or that specifying the return type as a `JavaValue::Object` requires a `String` allocation that's unused for making the method call.

I figured I'd add a missing changelog entry and minimal docs for the added `ReturnType` enum and submit a PR on behalf of @georgwi, since it didn't look like there was any push back against the idea, just an admission that the project isn't very actively maintained currently.

This updates the lower-level `call_static_method_unchecked` and `call_method_unchecked` APIs to take an array of `jvalue`s directly to avoid needing to collect `JValue` enum values into `jvalue` unions before calling into Java.

This also introduced a `ReturnType` that avoids the cost of allocating a `String` for the class name that's associated with `JavaType::Object(_)`.

### Definition of Done

- [x] There are no TODOs left in the code
- [x] Change is covered by [automated tests](https://github.com/jni-rs/jni-rs/blob/master/CONTRIBUTING.md#tests)
- [x] The [coding guidelines](https://github.com/jni-rs/jni-rs/blob/master/CONTRIBUTING.md#the-code-style) are followed
- [x] Public API has documentation
- [x] User-visible changes are mentioned in the Changelog
- [x] The continuous integration build passes