@rib 2022-10-14
Including the benchmark from https://github.com/jni-rs/jni-rs/pull/361 these are the benchmark results I see before and after this set of changes:

# _unchecked api call overhead
# Before
```
test tests::jni_call_object_hash_method_unchecked             ... bench:         177 ns/iter (+/- 16)
test tests::jni_call_static_abs_method_unchecked_jclass       ... bench:         247 ns/iter (+/- 12)
test tests::jni_call_static_date_time_method_unchecked_jclass ... bench:         468 ns/iter (+/- 47)
```
## After
```
test tests::jni_call_object_hash_method_unchecked             ... bench:         168 ns/iter (+/- 3)
test tests::jni_call_static_abs_method_unchecked_jclass       ... bench:         168 ns/iter (+/- 9)
test tests::jni_call_static_date_time_method_unchecked_jclass ... bench:         293 ns/iter (+/- 39)
```

As expected there's a particularly large improvement for the new `LocalDateTime::of` test which previously had to pass a `JValue::Object(String)` with a heap allocated string for every call

# All benchmarks
## Before
```
test tests::jni_call_object_hash_method_safe                  ... bench:       1,653 ns/iter (+/- 334)
test tests::jni_call_object_hash_method_unchecked             ... bench:         177 ns/iter (+/- 16)
test tests::jni_call_static_abs_method_safe                   ... bench:       2,460 ns/iter (+/- 352)
test tests::jni_call_static_abs_method_unchecked_jclass       ... bench:         247 ns/iter (+/- 12)
test tests::jni_call_static_abs_method_unchecked_str          ... bench:       1,009 ns/iter (+/- 117)
test tests::jni_call_static_date_time_method_safe             ... bench:       4,186 ns/iter (+/- 167)
test tests::jni_call_static_date_time_method_unchecked_jclass ... bench:         468 ns/iter (+/- 47)
test tests::jni_check_exception                               ... bench:          13 ns/iter (+/- 4)
test tests::jni_get_java_vm                                   ... bench:           5 ns/iter (+/- 0)
test tests::jni_new_global_ref                                ... bench:         186 ns/iter (+/- 4)
test tests::jni_new_object_by_id_jclass                       ... bench:         246 ns/iter (+/- 11)
test tests::jni_new_object_by_id_str                          ... bench:       1,112 ns/iter (+/- 64)
test tests::jni_new_object_jclass                             ... bench:       1,517 ns/iter (+/- 152)
test tests::jni_new_object_str                                ... bench:       2,357 ns/iter (+/- 226)
test tests::jni_noop_with_local_frame                         ... bench:          37 ns/iter (+/- 2)
test tests::jvm_noop_attach_detach_native_thread              ... bench:      49,981 ns/iter (+/- 4,659)
test tests::native_arc                                        ... bench:          11 ns/iter (+/- 1)
test tests::native_call_function                              ... bench:           1 ns/iter (+/- 0)
test tests::native_rc                                         ... bench:           3 ns/iter (+/- 0)
```

## After
```
test tests::jni_call_object_hash_method_safe                  ... bench:       1,500 ns/iter (+/- 258)
test tests::jni_call_object_hash_method_unchecked             ... bench:         168 ns/iter (+/- 3)
test tests::jni_call_static_abs_method_safe                   ... bench:       2,393 ns/iter (+/- 260)
test tests::jni_call_static_abs_method_unchecked_jclass       ... bench:         168 ns/iter (+/- 9)
test tests::jni_call_static_abs_method_unchecked_str          ... bench:         979 ns/iter (+/- 54)
test tests::jni_call_static_date_time_method_safe             ... bench:       3,813 ns/iter (+/- 310)
test tests::jni_call_static_date_time_method_unchecked_jclass ... bench:         293 ns/iter (+/- 39)
test tests::jni_check_exception                               ... bench:          12 ns/iter (+/- 0)
test tests::jni_get_java_vm                                   ... bench:           5 ns/iter (+/- 0)
test tests::jni_new_global_ref                                ... bench:         173 ns/iter (+/- 14)
test tests::jni_new_object_by_id_jclass                       ... bench:         255 ns/iter (+/- 17)
test tests::jni_new_object_by_id_str                          ... bench:       1,116 ns/iter (+/- 108)
test tests::jni_new_object_jclass                             ... bench:       1,398 ns/iter (+/- 117)
test tests::jni_new_object_str                                ... bench:       2,236 ns/iter (+/- 302)
test tests::jni_noop_with_local_frame                         ... bench:          38 ns/iter (+/- 1)
test tests::jvm_noop_attach_detach_native_thread              ... bench:      44,509 ns/iter (+/- 5,783)
test tests::native_arc                                        ... bench:          11 ns/iter (+/- 0)
test tests::native_call_function                              ... bench:           1 ns/iter (+/- 0)
test tests::native_rc                                         ... bench:           3 ns/iter (+/- 0)
```